Take collision logic out of WanderingCommander
